Add util to extract subtitle language code Refactor utility methods into a dedicated module.
