+++ media_manager/indexer/service.py
@@ -7,7 +7,7 @@ from media_manager.indexer.indexers.prowlarr import Prowlarr
 from media_manager.indexer.repository import IndexerRepository
 from media_manager.indexer.schemas import IndexerQueryResult, IndexerQueryResultId
 from media_manager.movies.schemas import Movie
-from media_manager.torrent.utils import remove_special_chars_and_parentheses
+from media_manager.utils.files import remove_special_chars_and_parentheses
 from media_manager.tv.schemas import Show
 
 log = logging.getLogger(__name__)
